requirements.txt
torch>=2.0
torchvision>=0.15
pennylane>=0.38
qiskit>=1.3
qiskit-aer>=0.17
qiskit-machine-learning>=0.8
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
codecarbon>=2.3
